+++ CHANGELOG.md
@@ -11,12 +11,14 @@
 * Renaming the `index` function as `row_index` for clarity.
 * `xan agg -C/--along-columns & -M/--along-matrix` & `xan groupby -C/--along-columns & -M/--along-matrix` will not map current column index to the result of the `index()` function. The `col_index()` can be now used instead for this very purpose.
 * `xan window -g/--groupby` does not require the file to be sorted anymore. This means using `-g/--groupby` will now require the whole file to be buffered into memory by the command. The old behavior can still be used through the `-S/--sorted` flag, thus aligning the `xan window` command with the rest of the tool.
+* `row_index` will now error if the expression has no concept of row index, instead of returning nothing.
 
 *Features*
 
 * Adding `xan grep -P/--add-pattern, --patterns, --pattern-column`.
 * Adding `xan parallel --dont-chunk`.
-* Adding nullary `col_index` & `header()` variants, to work with expression applied in series to multiple columns at once.
+* Adding nullary `col`, `col_index` & `header()` variants, to work with expression applied in series to multiple columns at once.
+* Adding `prev_col` & `next_col` functions.
 * Adding `xan filter -B/--before-context & -A/--after-context`.
 * Adding `xan window -O/--overwrite`.
 * Adding `xan map -C/--along-columns`.